% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{ancestorGraph}
\alias{ancestorGraph}
\title{Ancestor Graph}
\usage{
ancestorGraph(x, v = NULL)
}
\arguments{
\item{x}{the input graph.}

\item{v}{variable names.}
}
\description{
Creates the induced subgraph containing only the vertices
in \code{v}, their ancestors, and the edges between them. All
other vertices and edges are discarded.
}

