% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{ancestors}
\alias{ancestors}
\title{Get Graphical Ancestors}
\usage{
ancestors(x, v)
}
\arguments{
\item{x}{the input graph.}

\item{v}{character, name of a variable.}
}
\description{
Finds all variables from which \eqn{v} is reachable via a directed path. 
By definition, this includes \eqn{v} itself.
}

