% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{as.dagitty.data.frame}
\alias{as.dagitty.data.frame}
\title{Convert Lavaan Model to Dagitty Graph}
\usage{
as.dagitty.data.frame(x)
}
\arguments{
\item{x}{data frame, lavaan parameter table such as returned by
\code{\link[lavaan]{lavaanify}}.}
}
\description{
Convert Lavaan Model to Dagitty Graph
}

