% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.dagitty}
\alias{as.dagitty}
\title{Convert to DAGitty object}
\usage{
as.dagitty(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{further arguments passed on to methods.}
}
\description{
Converts its argument to a DAGitty object, if possible.
}
