% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{canonicalGraph}
\alias{canonicalGraph}
\title{Replace Bidirected Edges by Latent Variables}
\usage{
canonicalGraph(x)
}
\arguments{
\item{x}{the input graph.}
}
\description{
Replace Bidirected Edges by Latent Variables
}

