% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{children}
\alias{children}
\title{Get Graphical Children}
\usage{
children(x, v)
}
\arguments{
\item{x}{the input graph.}

\item{v}{character, name of a variable.}
}
\description{
Finds all variables \eqn{w} that are connected to \eqn{v}
by an edge \eqn{v} -> \eqn{w}.
}

