% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{completeDAG}
\alias{completeDAG}
\title{Generate Complete DAG}
\usage{
completeDAG(x)
}
\arguments{
\item{x}{variable names. Can also be a positive integer, in which case 
the variables will be called  x1,...,xN.}
}
\description{
Generates a complete DAG on the given variable names. The order
in which the variables are given corresponds to the topological ordering
of the DAG. Returns a named list.
}
