% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{convert}
\alias{convert}
\title{Convert from DAGitty object to other graph types}
\usage{
convert(x, to, ...)
}
\arguments{
\item{x}{a \code{dagitty} object or a character string.}

\item{to}{destination format, currently one of "dagitty", "tikz", "lavaan", "bnlearn", or "causaleffect".}

\item{...}{further arguments passed on to methods (currently unused)}
}
\description{
Converts its argument from a DAGitty object (or character string describing it)
to another package's format, if possible.
}
