% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{coordinates}
\alias{coordinates}
\title{Plot Coordinates of Variables in Graph}
\usage{
coordinates(x)
}
\arguments{
\item{x}{the input graph.}
}
\value{
A list with components \code{x} and \code{y}, giving relative coordinates
for each variable.
}
\description{
The DAGitty syntax allows specification of plot coordinates for each variable in a
graph. This function extracts these plot coordinates from the graph description in a
\code{dagitty} object. Note that the coordinate system is undefined, typically one
needs to compute the bounding box before plotting the graph.
}
\examples{
## Plot localization of each node in the Shier example
plot( coordinates( getExample("Shrier") ) )
}
\seealso{
Function \link{layout} for automtically generating layout coordinates, and function
\link{plot.dagitty} for plotting graphs.
}

