% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{dagitty}
\alias{dagitty}
\title{Parse Dagitty Graph}
\usage{
dagitty(x)
}
\arguments{
\item{x}{character, string describing a graphical model in dagitty syntax.}
}
\description{
Parse Dagitty Graph
}

