% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{dconnected}
\alias{dconnected}
\alias{dseparated}
\title{d-Separation}
\usage{
dconnected(x, X, Y = list(), Z = list())

dseparated(x, X, Y = list(), Z = list())
}
\arguments{
\item{x}{the input graph.}

\item{X}{vector of variable names.}

\item{Y}{vector of variable names.}

\item{Z}{vector of variable names.}
}
\description{
d-Separation
}

