% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{descendants}
\alias{descendants}
\title{Get Graphical Descendants}
\usage{
descendants(x, v)
}
\arguments{
\item{x}{the input graph.}

\item{v}{character, name of a variable.}
}
\description{
Finds all variables that are reachable from \eqn{v} via a directed path. By definition,
this includes \eqn{v} itself.
}

