% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{downloadGraph}
\alias{downloadGraph}
\title{Load Graph from dagitty.net}
\usage{
downloadGraph(x = "dagitty.net/mz-Tuw9")
}
\arguments{
\item{x}{dagitty model URL.}
}
\description{
Load Graph from dagitty.net
}

