% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{exogenousVariables}
\alias{exogenousVariables}
\title{Retrieve Exogenous Variables}
\usage{
exogenousVariables(x)
}
\arguments{
\item{x}{the input graph, of any type.}
}
\description{
Returns the names of all variables that have no directed arrow pointing to them.
Note that this does not preclude variables connected to bidirected arrows.
}
