% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{impliedConditionalIndependencies}
\alias{impliedConditionalIndependencies}
\title{List Conditional Indpendencies Implied by Graphical Model}
\usage{
impliedConditionalIndependencies(x, max.results = 100)
}
\arguments{
\item{x}{the input graph.}

\item{max.results}{integer. The listing of conditional independencies is stopped once
this many results have been found. Use \code{Inf} to generate them all.}
}
\description{
List Conditional Indpendencies Implied by Graphical Model
}

