% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{instrumentalVariables}
\alias{instrumentalVariables}
\title{Find (Conditional) Instrumental Variables to Estimate Causal Effects}
\usage{
instrumentalVariables(x, exposure = NULL, outcome = NULL)
}
\arguments{
\item{x}{the input graph.}

\item{exposure}{name of the exposure variable. If not given (default), then the 
 exposure variable is supposed to be defined in the graph itself. Only a single exposure
variable is supported.}

\item{outcome}{name of the outcome variable, also taken from the graph if not given.
Only a single outcome variable is supported.}
}
\description{
Find (Conditional) Instrumental Variables to Estimate Causal Effects
}

