% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.R
\name{is.dagitty}
\alias{is.dagitty}
\title{Test for Graph Class}
\usage{
is.dagitty(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
A function to check whether an object has class \code{dagitty}.
}
