% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{localTests}
\alias{localTests}
\title{Perform Local Tests of Structural Equation Model}
\usage{
localTests(x, data, type = "tetrads", R = 50)
}
\arguments{
\item{x}{the model.}

\item{data}{the data.}

\item{type}{character indicating which kind of local
test to perform.}

\item{R}{how many bootstrap replicates for estimating tetrad
distribution.}
}
\description{
Perform Local Tests of Structural Equation Model
}

