% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{measurementPart}
\alias{measurementPart}
\title{Extract Measurement Part from Structural Equation Model}
\usage{
measurementPart(x)
}
\arguments{
\item{x}{the input graph, a DAG.}
}
\description{
Removes all edges between latent variables, then removes any
latent variables without adjacent edges, then returns the graph.
}
\details{
Assumes that x is a graph where there are edges between 
the latent variables, between the observed variables, and 
from latent to observed variables, but no edge between
a latent L and an observed X may have an arrowhead at L.
}
