% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{parents}
\alias{parents}
\title{Get Graphical Parents}
\usage{
parents(x, v)
}
\arguments{
\item{x}{the input graph.}

\item{v}{character, name of a variable.}
}
\description{
Finds all variables \eqn{w} to which \eqn{v} is connected
by an edge \eqn{w} -> \eqn{v}.
}

