% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{plot.dagitty}
\alias{plot.dagitty}
\title{Plotting Graph}
\usage{
\method{plot}{dagitty}(x, ...)
}
\arguments{
\item{x}{the input graph.}

\item{...}{not used.}
}
\description{
A simple plot method to quickly visualize a graph. This is intended mainly for
validation purposes and is not yet meant to become a full-fledged graph drawing
function.
}

