% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{randomDAG}
\alias{randomDAG}
\title{Generate DAG at Random}
\usage{
randomDAG(N, p)
}
\arguments{
\item{N}{desired number of variables.}

\item{p}{connectivity parameter, a number between 0 and 1.}
}
\description{
Generates a random DAG with N variables called x1,...,xN. For each
pair of variables xi,xj with i<j, an edge i->j will be present with
probability p.
}

