% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/dagitty.r
\name{setVariableStatus}
\alias{setVariableStatus}
\title{Change Status of Variables in Graph}
\usage{
setVariableStatus(x, var.names, status)
}
\arguments{
\item{x}{the input graph.}

\item{var.names}{names of variables to change.}

\item{status}{character, one of "exposure" or "outcome".}
}
\description{
First removes the given status from all variables in the graph that had it, and then
gives it to the given variables.
For instance, if  \code{status="exposure"}  and \code{var.names="X"} are given, then
\code{X} will be the only exposure in the resulting graph.
}

