% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{structuralPart}
\alias{structuralPart}
\title{Extract Structural Part from Structural Equation Model}
\usage{
structuralPart(x)
}
\arguments{
\item{x}{the input graph, a DAG.}
}
\description{
Removes all observed variables from the input graph.
}
\details{
Assumes that x is a graph where there are edges between 
the latent variables, between the observed variables, and 
from latent to observed variables, but no edge between
a latent L and an observed X may have an arrowhead at L.
}

