% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.R
\name{topologicalOrdering}
\alias{topologicalOrdering}
\title{Get Topological Ordering of DAG}
\usage{
topologicalOrdering(x)
}
\arguments{
\item{x}{the input graph, a DAG}
}
\description{
Computes a topological ordering of the nodes, i.e., a number for each node
such that every node's number is smaller than the one of all its descendants.
Bidirected edges (<->) are ignored.
}
