% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{vanishingTetrads}
\alias{vanishingTetrads}
\title{List Vanishing Tetrads Implied by Gaussian Graphical Model}
\usage{
vanishingTetrads(x)
}
\arguments{
\item{x}{the input graph.}
}
\description{
List Vanishing Tetrads Implied by Gaussian Graphical Model
}

