% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{adjustmentSets}
\alias{adjustmentSets}
\title{Find Adjustment Sets to Estimate Causal Effects}
\usage{
adjustmentSets(x, exposure = NULL, outcome = NULL, effect = "total")
}
\arguments{
\item{x}{the input graph.}

\item{exposure}{name(s) of the exposure variable(s). If not given (default), then the 
exposure variables are supposed to be defined in the graph itself.}

\item{outcome}{name(s) of the outcome variable(s), also taken from the graph if 
not given.}

\item{effect}{which effect is to be identified. If \code{effect="total"}, then the
total effect is to be identified, and the adjustment criterion by van der Zander et 
al (2014), an extension of Pearl's back-door criterion, is used. Otherwise, if 
\code{effect="direct"}, then the average direct effect is to be identified, and Pearl's
single-door criterion is used (Pearl, 2009). In a structural equation model (Gaussian
graphical model), direct effects are simply the path coefficients.}
}
\description{
Find Adjustment Sets to Estimate Causal Effects
}

