% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{dagitty}
\alias{dagitty}
\title{Parse Dagitty Graph}
\usage{
dagitty(x)
}
\arguments{
\item{x}{character, string describing a graphical model in dagitty syntax.}
}
\description{
Constructs a \code{dagitty} graph object from a textual description.
}

