% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.R
\name{downloadGraph}
\alias{downloadGraph}
\title{Load Graph from dagitty.net}
\usage{
downloadGraph(x = "dagitty.net/mz-Tuw9")
}
\arguments{
\item{x}{dagitty model URL.}
}
\description{
Downloads a graph that has been built and stored online using the dagitty.net GUI.
Users who store graphs online will receive a unique URL for their graph, which
can be fed into this function to continue working with the graph in R.
}
