% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{isAdjustmentSet}
\alias{isAdjustmentSet}
\title{Adjustment Criterion}
\usage{
isAdjustmentSet(x, Z, exposure = NULL, outcome = NULL)
}
\arguments{
\item{x}{the input graph.}

\item{Z}{vector of variable names.}

\item{exposure}{name(s) of the exposure variable(s). If not given (default), then the 
exposure variables are supposed to be defined in the graph itself.}

\item{outcome}{name(s) of the outcome variable(s), also taken from the graph if 
not given.}
}
\description{
Test whether a set fulfills the adjustment criterion, that means,
it removes all confounding bias when estimating a *total* effect.
This is an extension of Pearl's 
Back-door criterion which is complete in the sense that either a set
fulfills this criterion, or it does not remove all confounding bias.
}

