% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{localTests}
\alias{localTests}
\title{Perform Local Tests of Structural Equation Model}
\usage{
localTests(x, data = NULL, type = "tetrads", sample.cov = NULL,
  sample.nobs = NULL, conf.level = 0.95, R = NULL)
}
\arguments{
\item{x}{the model.}

\item{data}{matrix or data frame containing the data.}

\item{type}{character indicating which kind of local
test to perform. Supported values are \code{"cis"},
\code{"tetrads"} and \code{"tetrads.type"}, where "type" is one of the items of the 
tetrad typology, e.g. \code{"tetrads.within"} (see \code{\link{vanishingTetrads}}).}

\item{sample.cov}{the sample covariance matrix; ignored if \code{data} is supplied.
Either \code{data} or \code{sample.cov} and \code{sample.nobs} must be supplied.}

\item{sample.nobs}{number of observations; ignored if \code{data} is supplied.}

\item{conf.level}{determines the size of confidence intervals for test
statistics.}

\item{R}{how many bootstrap replicates for estimating confidence
intervals. If \code{NULL}, then confidence intervals are based on normal
approximation. For tetrads, the normal approximation is only valid in 
large samples even if the data are normally distributed.}
}
\description{
Perform Local Tests of Structural Equation Model
}

