% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{localTests}
\alias{localTests}
\title{Perform Local Tests of Structural Equation Model}
\usage{
localTests(x, data, type = "tetrads", conf.level = 0.95, R = NA)
}
\arguments{
\item{x}{the model.}

\item{data}{the data.}

\item{type}{character indicating which kind of local
test to perform. Supported values are \code{"tetrads"} and 
\code{"cis"}.}

\item{conf.level}{determines the size of confidence intervals for test
statistics.}

\item{R}{how many bootstrap replicates for estimating confidence
intervals. If \code{NA}, then confidence intervals are based on normal
approximation. For tetrads, the normal approximation is only valid in 
large samples even if the data are normally distributed.}
}
\description{
Perform Local Tests of Structural Equation Model
}

