% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{plot.dagitty}
\alias{plot.dagitty}
\title{Plot Graph}
\usage{
\method{plot}{dagitty}(
  x,
  abbreviate.names = FALSE,
  show.coefficients = FALSE,
  adjust.coefficients = NA,
  ...
)
}
\arguments{
\item{x}{the input graph, a DAG, MAG, or PDAG.}

\item{abbreviate.names}{logical. Whether to abbreviate variable names.}

\item{show.coefficients}{logical. Whether to plot coefficients defined in the graph syntax
on the edges.}

\item{adjust.coefficients}{numerical. Adjustment for coefficient labels; the distance between 
the edge labels and the midpoint of the edge can be controlled using this paramer. 
Can also be a vector of 2 numbers for separate horizontal and vertical adjustment. NA means
no adjustment (default).}

\item{...}{not used.}
}
\description{
A simple plot method to quickly visualize a graph. This is intended mainly for 
simple visualization purposes and not as a full-fledged graph drawing
function.
}
