% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty.r
\name{vanishingTetrads}
\alias{vanishingTetrads}
\title{List Vanishing Tetrads Implied by Gaussian Graphical Model}
\usage{
vanishingTetrads(x, type = NA)
}
\arguments{
\item{x}{the input graph.}

\item{type}{restrict output to one level of Kenny's tetrad typology.
Possible values are "within" (homogeneity within constructs; all four
variables have the same parents), "between" (homogeneity between constructs;
two pairs of variables each sharing one parent) 
and "epistemic" (consistency of epistemic correlations; three variables have
the same parent). By default, all tetrads are listed.}
}
\value{
a data frame with four columns, where each row of the form
i,j,k,l means that the tetrad Cov(i,j)Cov(k,l) - Cov(i,k)Cov(j,l) vanishes
(is equal to 0) according to the model.
}
\description{
List Vanishing Tetrads Implied by Gaussian Graphical Model
}

